Add files via upload
After both inspecting the example site and other examples on MDN and Stak overflow, this is what I've produced. 

Biggest challenge was re-tailoring the header. It was initially set up for a desktop and had nav links but recognised I needed to restructure and create a simpler design if I wanted to comply with Tablet and Mobile viewports.